# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /agent/chat:
    post:
      operationId: agentChat
      tags: [agent]
      summary: Proxy streaming AI chat to Anthropic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentChatRequest"
      responses:
        "200":
          description: SSE stream of text deltas
          content:
            text/event-stream:
              schema:
                type: string
components:
  schemas:
...
    AgentChatRequest:
      type: object
      properties:
        messages:
          type: array
          items:
            type: object
            properties:
              role:
                type: string
                enum: [user, assistant]
              content:
                type: string
            required: [role, content]
        system:
          type: string
        max_tokens:
          type: integer
      required:
        - messages
...